requirements.txt
discord
requests
PyNaCl
python-dotenv
youtube-dl
ffmpeg